Update CartMapper.java
20220724 리턴 타입 변경

+++ Yumyum/src/main/java/com/yum/mapper/CartMapper.java
@@ -29,7 +29,7 @@ public interface CartMapper {
 			, @Param("branchNum") Long branchNum);
 	
 	// 선택한 지점에 대한 장바구니 확인
-	public Long countTotalCart(
+	public int countTotalCart(
 			@Param("userNum") Long userNum
 			, @Param("branchNum") Long branchNum);
 }
